Screen colour and tool functionality
The enum defining the colours has been changed. It now changes the text colour to the same colour as the background, meaning there are no more dots visible on the screen (The dots are needed to set the width correctly)

The "Fill square" function for the tools has been implemented. You can either specify a single square to fill from coordinates, or a vector of coordinates to fill. The next step is creating functions for the line and fill tools that create a vector to pass to this function.