Main class setup
Up until this point, commands had to be manually coded into the main.cpp file to get things to happen on-screen. I have begun setting up the main class, which will allow the user to input their own commands on the fly.

I have also tidied parts of the code up. For example, a function in a command required itself to be passed as a parameter, which wasn't necessary