Open & new features
Opening saved files has been fixed. The error was due to the program deleting the existing file multiple times.

Creating a new file was also fixed and reworked. It was previously handled by the screen, which does not make sense. It is now handled by the main paint class, and this caused the screen to lose quite a bit of bloat.